# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/resources": {
      "get": {
        "description": "支持按 id/关键字/分类分页查询资源列表。",
        "operationId": "ResourceController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数，最大 100",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "资源 id（单 id 场景只用 id）",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "标题关键字",
            "schema": {
              "example": "演示",
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "分类",
            "schema": {
              "example": "movie",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "资源列表查询",
        "tags": [
          "资源模块"
        ]
      }
    },
    "/resource-detail": {
      "get": {
        "description": "单对象操作统一使用 query 参数 id，不使用路径参数。",
        "operationId": "ResourceController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "资源主键 id（UUID）",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "资源详情查询",
        "tags": [
          "资源模块"
        ]
      }
    },
    "/resource-episodes": {
      "get": {
        "description": "按资源 id 查询分集列表，单对象参数统一使用 id。",
        "operationId": "ResourceController_episodes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "资源主键 id（UUID）",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpisodeDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "资源分集查询",
        "tags": [
          "资源模块"
        ]
      }
    },
    "/crawl-job/create": {
      "post": {
        "description": "创建一个资源站目录元数据采集任务。",
        "operationId": "CrawlJobController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCrawlJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlJobResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建采集任务",
        "tags": [
          "采集任务模块"
        ]
      }
    },
    "/crawl-job/detail": {
      "get": {
        "description": "单对象操作统一使用 query 参数 id。",
        "operationId": "CrawlJobController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "任务主键 id（UUID）",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlJobResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "采集任务详情",
        "tags": [
          "采集任务模块"
        ]
      }
    },
    "/parser-sources": {
      "get": {
        "description": "按 system/user 范围查询解析源配置列表。",
        "operationId": "ParserSourceController_list",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "查询范围：system 或 user",
            "schema": {
              "example": "system",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ParserSourceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "解析源列表",
        "tags": [
          "解析源模块"
        ]
      }
    },
    "/parser-source/create": {
      "post": {
        "description": "新增用户自定义解析源。",
        "operationId": "ParserSourceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParserSourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParserSourceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建解析源",
        "tags": [
          "解析源模块"
        ]
      }
    },
    "/parser-source/update": {
      "post": {
        "description": "单对象更新统一使用 body.id。",
        "operationId": "ParserSourceController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParserSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParserSourceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新解析源",
        "tags": [
          "解析源模块"
        ]
      }
    },
...
    "/knowledge/agent-note/create": {
      "post": {
        "description": "写入任务知识，用于后续复用，减少重复分析。",
        "operationId": "KnowledgeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentNoteResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建知识记录",
        "tags": [
          "知识沉淀模块"
        ]
      }
    },
    "/knowledge/agent-note/search": {
      "get": {
        "description": "按关键字或 taskKey 检索历史知识。",
        "operationId": "KnowledgeController_search",
        "parameters": [
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字搜索",
            "schema": {
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "taskKey",
            "required": false,
            "in": "query",
            "description": "任务标识过滤",
            "schema": {
              "example": "video-resource-v1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentNoteResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "搜索知识记录",
        "tags": [
          "知识沉淀模块"
        ]
      }
    },
    "/knowledge/agent-note/detail": {
      "post": {
        "description": "单对象查询统一使用 body.id。",
        "operationId": "KnowledgeController_detail",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "仅作规则提示：单对象参数统一为 id",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentNoteDetailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentNoteResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "知识记录详情",
        "tags": [
          "知识沉淀模块"
        ]
      }
    },
...
  },
  "info": {
    "title": "Isle Flix 服务端接口文档",
    "description": "视频资源聚合服务接口文档（中文）。单对象操作统一使用 id，禁止路径参数。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "先调用 /auth/login 获取 accessToken，再在此处填入。"
      }
    },
    "schemas": {
...
      "ResourceItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "资源主键 id",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "description": "资源标题",
            "example": "Demo Video A"
          },
          "category": {
            "type": "string",
            "description": "资源分类",
            "example": "movie"
          },
          "siteKey": {
            "type": "string",
            "description": "来源站点标识",
            "example": "yszzq"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "siteKey"
        ]
      },
      "ResourceListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "列表数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "总条数",
            "example": 1
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "EpisodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "分集主键 id",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "resourceId": {
            "type": "string",
            "description": "所属资源 id",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "分集名称",
            "example": "Episode 1"
          },
          "playPageUrl": {
            "type": "string",
            "description": "播放页地址",
            "example": "https://www.yszzq.com/ziyuan/"
          }
        },
        "required": [
          "id",
          "resourceId",
          "name",
          "playPageUrl"
        ]
      },
      "CreateCrawlJobDto": {
        "type": "object",
        "properties": {
          "siteKey": {
            "type": "string",
            "description": "站点标识",
            "example": "yszzq"
          },
          "category": {
            "type": "string",
            "description": "可选分类过滤",
            "example": "movie"
          }
        },
        "required": [
          "siteKey"
        ]
      },
      "CrawlJobResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "任务主键 id",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "siteKey": {
            "type": "string",
            "description": "站点标识",
            "example": "yszzq"
          },
          "category": {
            "type": "string",
            "description": "任务分类",
            "example": "movie"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "example": "queued"
          }
        },
        "required": [
          "id",
          "siteKey",
          "status"
        ]
      },
      "ParserSourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "解析源主键 id",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "解析源名称",
            "example": "system-default"
          },
          "endpoint": {
            "type": "string",
            "description": "解析服务地址",
            "example": "https://parser.example.com"
          },
          "enabled": {
            "type": "boolean",
            "description": "是否启用",
            "example": true
          },
          "ownerType": {
            "type": "string",
            "description": "归属类型",
            "example": "system"
          }
        },
        "required": [
          "id",
          "name",
          "endpoint",
          "enabled",
          "ownerType"
        ]
      },
      "CreateParserSourceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "解析源名称",
            "example": "系统默认解析源"
          },
          "endpoint": {
            "type": "string",
            "description": "解析服务地址",
            "example": "https://parser.example.com"
          }
        },
        "required": [
          "name",
          "endpoint"
        ]
      },
      "UpdateParserSourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "解析源主键 id（单 id 场景统一使用 id）",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "解析源名称",
            "example": "用户自定义解析源"
          },
          "endpoint": {
            "type": "string",
            "description": "解析服务地址",
            "example": "https://example.com/parser"
          },
          "enabled": {
            "type": "boolean",
            "description": "是否启用",
            "example": true
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CreateAgentNoteDto": {
        "type": "object",
        "properties": {
          "taskKey": {
            "type": "string",
            "description": "任务标识",
            "example": "video-resource-v1"
          },
          "content": {
            "type": "string",
            "description": "任务知识内容",
            "example": "统一约定单对象参数使用 id。"
          }
        },
        "required": [
          "taskKey",
          "content"
        ]
      },
      "AgentNoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "知识记录主键 id",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "taskKey": {
            "type": "string",
            "description": "任务标识",
            "example": "video-resource-v1"
          },
          "content": {
            "type": "string",
            "description": "知识内容",
            "example": "统一约定单对象参数使用 id。"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2026-05-14T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "taskKey",
          "content",
          "createdAt"
        ]
      },
      "AgentNoteDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "知识记录主键 id（单 id 场景统一使用 id）",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}